+++ api/test/utils.py
@@ -4,8 +4,12 @@ import json
 import os
 from datetime import datetime
 from bson.objectid import ObjectId
+from celery.task.base import Task
 
 from api import app
+
+
+MODEL_ID = '519318e6106a6c0df349bc0b'
 
 
 class BaseTestCase(unittest.TestCase):
@@ -31,6 +35,7 @@ class BaseTestCase(unittest.TestCase):
 
     @classmethod
     def fixtures_load(cls):
+        from bson.objectid import ObjectId
         for fixture in cls.FIXTURES:
             data = _load_fixture_data(fixture)
             for collection_name, documents in data.iteritems():
@@ -125,6 +130,40 @@ class BaseTestCase(unittest.TestCase):
             obj = self.Model.find_one({'_id': ObjectId(_id)})
             return resp, obj
         return resp
+
+
+class CeleryTestCaseBase(BaseTestCase):
+
+    def setUp(self):
+        super(CeleryTestCaseBase, self).setUp()
+        self.applied_tasks = []
+
+        self.task_apply_async_orig = Task.apply_async
+
+        @classmethod
+        def new_apply_async(task_class, args=None, kwargs=None, **options):
+            self.handle_apply_async(task_class, args, kwargs, **options)
+
+        # monkey patch the regular apply_sync with our method
+        Task.apply_async = new_apply_async
+
+    def tearDown(self):
+        super(CeleryTestCaseBase, self).tearDown()
+
+        # Reset the monkey patch to the original method
+        Task.apply_async = self.task_apply_async_orig
+
+    def handle_apply_async(self, task_class, args=None, kwargs=None, **options):
+        self.applied_tasks.append((task_class, tuple(args), kwargs))
+
+    def assert_task_sent(self, task_class, *args, **kwargs):
+        was_sent = any(task_class == task[0] and args == task[1] and kwargs == task[2]
+                       for task in self.applied_tasks)
+        self.assertTrue(was_sent, 'Task not called w/class %s and args %s' % (task_class, args))
+
+    def assert_task_not_sent(self, task_class):
+        was_sent = any(task_class == task[0] for task in self.applied_tasks)
+        self.assertFalse(was_sent, 'Task was not expected to be called, but was.  Applied tasks: %s' %                 self.applied_tasks)
 
 
 def _get_collection(name):