+++ manage.py
@@ -4,10 +4,6 @@ from flask.ext.script import Manager, Command, Shell
 
 from api import app
 
-import gevent.monkey
-from gevent.pywsgi import WSGIServer
-gevent.monkey.patch_all()
-
 
 class Celeryd(Command):
     """Runs a Celery worker node."""
@@ -26,6 +22,9 @@ class Run(Command):
     """Runs a Celery Flower worker node."""
 
     def run(self, **kwargs):
+        import gevent.monkey
+        from gevent.pywsgi import WSGIServer
+        gevent.monkey.patch_all()
         http_server = WSGIServer(('', 5000), app)
         http_server.serve_forever()
 
@@ -33,9 +32,17 @@ class Migrate(Command):
     """Migrate"""
 
     def run(self, **kwargs):
-        from api.migrations import ModelMigration
-        from api.models import Model
-        ModelMigration(Model).migrate_all(app.db.Model.collection)
+        from api.migrations import ModelMigration, TestMigration
+        from api.models import Model, Test
+        #TestMigration(Test).migrate_all(app.db.Test.collection)
+        target = {'model_id': {'$exists': False}, 'model_name': {'$exists': True}}
+        
+        #ModelMigration(Model).migrate_all(app.db.Model.collection)
+        for doc in app.db.tests.find(target):
+            model_id = app.db.model.find_one({'name': doc['model_name']})['_id']
+            update = {'$set': {'model_id': model_id}}
+            print model_id
+            app.db.tests.update(target, update, multi=True, safe=True)
 
 
 def _make_context():